refactor: Mode type to types
